% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc-utils.R
\docType{data}
\name{.cdisc_lookup}
\alias{.cdisc_lookup}
\title{CDISC Variable Recognition and Utilities}
\format{
An object of class \code{list} of length 9.
}
\usage{
.cdisc_lookup
}
\description{
Functions to automatically detect and work with CDISC (Clinical Data Interchange
Standards Consortium) variable naming conventions commonly used in clinical trials.
}
\details{
These functions help identify standard CDISC variable names and provide
suggestions for proper formula construction in longitudinal clinical trial analysis.
}
\keyword{datasets}
