% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-themes.R
\name{get_publication_theme}
\alias{get_publication_theme}
\title{Get Publication Theme by Name}
\usage{
get_publication_theme(theme_name = "nature", ...)
}
\arguments{
\item{theme_name}{Character string specifying theme name.
Options: "nature", "science", "nejm", "fda", "default".}

\item{...}{Additional arguments passed to specific theme functions.}
}
\value{
A ggplot2 theme object.
}
\description{
Convenience function to get publication themes by name.
}
\examples{
theme_pub <- get_publication_theme("nature")
theme_reg <- get_publication_theme("fda", high_contrast = TRUE)

}
